+++ src/model/local_model_view/custom_model_details_dialog.rs
