+++ src/model/local_model_view/local_model_delete_confirmation_dialog.rs
