+++ src/model/local_model_view/local_model_view_helpers.rs
